Switches from CyclicBarrier to Phaser in MultiplexInputStream.
This will allow deregistration (through the InputStream's close() method) of consumers from the multiplexer. It should also fix issues with the time-limited barrier conditions but there are still issues in the code with that. Currently, a single slow thread (in a test I've not committed yet) simulated by a simple Thread.sleep is causing the entire collection to pause. Clearly I have issues with the manner in which I'm using awaitAdvanceInterruptibly() but I have yet to figure out what's wrong.